requirements.txt
asgiref==3.3.1
astroid==2.5.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.4.7
DateTime==4.3
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.7
django-ckeditor==6.0.0
django-countries==7.1
django-crispy-forms==1.11.2
django-etc==1.3.0
django-filter==2.4.0
django-fontawesome==1.0
django-fontawesome-5==1.0.18
django-heroku==0.3.1
django-hitcount==1.3.3
django-js-asset==1.2.2
django-mptt==0.12.0
django-view-breadcrumbs==2.1.1
django-widgets==0.1.16
genre==0.1.1
gitdb==4.0.7
GitPython==3.1.14
gunicorn==20.1.0
hierarchy==0.4.0
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
oauthlib==3.1.0
Pillow==8.1.2
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
pylint==2.7.3
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
pyuca==1.2
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
smmap==4.0.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
whitenoise==5.2.0
wrapt==1.12.1
zope.interface==5.4.0
